add:Initialization
Cree la inicializacion, cambiaron los imports y la estructura de los datos dado que dabla clavos por imports ciclicos, ademas se pueden meter varias variables
