+++ Textos/DeclaracionDeClases.txt
@@ -1,9 +1,11 @@
 class Perrito{
-    init(nombre){
+    init(nombre, edad, anios){
         this.nombre = nombre;
+        this.edad = edad;
+        this.siglo = anios;
     }
     caminar(){
         this.nombre = this.nombre + 1;
     }
 }
-var variable = new Perrito();
+var variable = new Perrito("raul", 20, 50);
